Making work of the login team code functionality
