+++ packages/client/src/containers/LandingPage/LandingPage.js
@@ -1,10 +1,35 @@
-import React from 'react';
+import React, { useState } from 'react';
+import { apiURL } from '../../apiURL';
 import Testimonials from '../../components/Login/Testimonials';
 import logo from './logo.svg';
 import { Typography, TextField, Button } from '@mui/material';
 import './LandingPage.Style.css';
 
 const LandingPageContainer = () => {
+  const [teamCode, setTeamCode] = useState('');
+  const [loginStatus, setLoginStatus] = useState(null);
+
+  const handleLogin = async (e) => {
+    e.preventDefault();
+    try {
+      const response = await fetch(`${apiURL()}/teams/login`, {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify({ team_code: teamCode }),
+      });
+
+      if (!response.ok) {
+        throw new Error('Login failed');
+      }
+
+      setLoginStatus('success');
+    } catch (error) {
+      setLoginStatus('failure');
+    }
+  };
+
   return (
     <div className="landing-page">
       <div className="testimonial-container">
@@ -22,24 +47,33 @@ const LandingPageContainer = () => {
           <Typography className="login-heading" variant="h5">
             Log in to Teams Check-in
           </Typography>
-          <div className="login-action">
-            <div className="login-text">
-              <Typography className="text-field" variant="body1">
-                Enter your team code
-              </Typography>
-              <TextField
-                className="team-code"
-                label="Your code"
-                type="password"
-                variant="outlined"
-                size="small"
-                required
-              />
-            </div>
-            <Button className="button" variant="contained" fullWidth>
+          <form className="login-text" onSubmit={handleLogin}>
+            <Typography className="text-field" variant="body1">
+              Enter your team code
+            </Typography>
+            <TextField
+              className="team-code"
+              label="Your code"
+              type="password"
+              variant="outlined"
+              size="small"
+              value={teamCode}
+              onChange={(e) => setTeamCode(e.target.value)}
+              required
+            />
+            <Button
+              type="submit"
+              className="button"
+              variant="contained"
+              fullWidth
+            >
               Login
             </Button>
-          </div>
+          </form>
+          {loginStatus === 'success' && <p>Login successful</p>}
+          {loginStatus === 'failure' && (
+            <p>Login unsuccessful. Please try again.</p>
+          )}
         </div>
       </div>
     </div>
